Modify behavior of enabling inline media
Previous behavior, when enabling inline media, was to show user a prompt that their IP address may be leaked. What is the point of a warning if a leak can occur while the user is reading the prompt because the feature is now enabled?

New behavior requires user to grant permission through the prompt to enable inline media. The method for this prompt has been relocated to TVCLogControllerInlineMediaService because it's related to inline media and I could not think of a better place.

Furthermore, the prompt has been configured so that clicking the button to open System Settings to configure a proxy does not dismiss the prompt.